Refactor CLI into package and introduce operations use cases
