+++ mlox/cli/commands/model.py
@@ -1,0 +1,1 @@
+from mlox.cli.app import model_list, model_deploy
