+++ mlox/cli/commands/project.py
@@ -1,0 +1,1 @@
+from mlox.cli.app import project_new
