+++ mlox/cli/commands/service.py
@@ -1,0 +1,1 @@
+from mlox.cli.app import service_list, service_add, service_setup, service_teardown, service_logs, service_configs_list
