+++ mlox/cli/rendering/table.py
@@ -1,0 +1,1 @@
+from mlox.cli.app import render_table
